Move twitter sign-in under Context
