+++ src/components/header/elements/connect-modal/ConnectModal.tsx
@@ -1,10 +1,6 @@
-import { getPublicKey } from '@noble/secp256k1';
 import classnames from 'classnames';
-import { useAtomValue } from 'jotai';
 import { useTranslation } from 'react-i18next';
-import { useCallback, useRef } from 'react';
-import { bytesToHex } from 'viem';
-import { useAccount, useChainId } from 'wagmi';
+import { useAccount } from 'wagmi';
 
 import { AccountBalanceWallet, CheckCircleOutline } from '@mui/icons-material';
 import { Button, DialogTitle, Typography } from '@mui/material';
@@ -13,8 +9,6 @@ import { Web3AuthConnectButton } from 'components/web3auth-connect-button/Web3Au
 import { WalletConnectButton } from 'components/wallet-connect-button/WalletConnectButton';
 import { Dialog } from 'components/dialog/Dialog';
 import { Separator } from 'components/separator/Separator';
-import { postSocialVerify } from 'network/referral';
-import { socialPKAtom, socialUserInfoAtom } from 'store/web3-auth.store';
 
 import styles from './ConnectModal.module.scss';
 
@@ -28,37 +22,6 @@ export const ConnectModal = ({ isOpen, onClose }: ConnectModalPropsI) => {
 
   const { isConnected } = useAccount();
 
-  const userInfo = useAtomValue(socialUserInfoAtom);
-  const socialPK = useAtomValue(socialPKAtom);
-
-  const verifyRef = useRef(false);
-
-  const chainId = useChainId();
-
-  const handleWeb3AuthSuccessConnect = useCallback(() => {
-    const verify = async () => {
-      if (!chainId || !userInfo?.idToken || verifyRef.current || !socialPK) {
-        return;
-      }
-      try {
-        verifyRef.current = true;
-        const pubKey = bytesToHex(getPublicKey(socialPK));
-        await postSocialVerify(chainId, userInfo.idToken, pubKey).catch((e) =>
-          console.log('POST /social-verify error', e)
-        );
-      } catch (error) {
-        console.log(error);
-      } finally {
-        verifyRef.current = false;
-      }
-    };
-    verify().then();
-  }, [chainId, socialPK, userInfo]);
-
-  const handleWeb3AuthErrorConnect = useCallback((error: string) => {
-    console.log(error);
-  }, []);
-
   return (
     <Dialog open={isOpen} onClose={onClose} className={styles.dialog}>
       {!isConnected && (
@@ -70,11 +33,7 @@ export const ConnectModal = ({ isOpen, onClose }: ConnectModalPropsI) => {
           <Separator />
           <div className={styles.dialogContent}>
             <div className={styles.actionButtonsContainer}>
-              <Web3AuthConnectButton
-                buttonClassName={styles.connectButton}
-                successCallback={handleWeb3AuthSuccessConnect}
-                errorCallback={handleWeb3AuthErrorConnect}
-              />
+              <Web3AuthConnectButton buttonClassName={styles.connectButton} />
               <div className={styles.orSeparator}>
                 <Separator />
                 <div className={styles.orTextHolder}>
